requirements.txt
flask
flask_cors
imagehash
Pillow
opencv-python-headless
python-magic
waitress
